refactor: make plugin-less as a standalone package
