+++ packages/core/package.json
@@ -71,7 +71,6 @@
     "http-compression": "1.0.20",
     "launch-editor-middleware": "^2.6.1",
     "fs-extra": "^11.2.0",
-    "less-loader": "12.2.0",
     "on-finished": "2.4.1",
     "open": "^8.4.0",
     "postcss-load-config": "5.1.0",
