+++ packages/core/tests/css.test.ts
@@ -6,7 +6,6 @@ import {
   normalizeCssLoaderOptions,
   pluginCss,
 } from '../src/plugins/css';
-import { pluginLess } from '../src/plugins/less';
 import { pluginSass } from '../src/plugins/sass';
 
 describe('normalizeCssLoaderOptions', () => {
@@ -170,70 +169,6 @@ describe('plugin-css injectStyles', () => {
 
     const bundlerConfigs = await rsbuild.initConfigs();
 
-    expect(bundlerConfigs[0]).toMatchSnapshot();
-  });
-});
-
-describe('plugin-less', () => {
-  it('should add less-loader', async () => {
-    const rsbuild = await createStubRsbuild({
-      plugins: [pluginLess()],
-      rsbuildConfig: {
-        tools: {
-          less: {},
-        },
-      },
-    });
-
-    const bundlerConfigs = await rsbuild.initConfigs();
-    expect(bundlerConfigs[0]).toMatchSnapshot();
-  });
-
-  it('should add less-loader and css-loader when injectStyles', async () => {
-    const rsbuild = await createStubRsbuild({
-      plugins: [pluginLess()],
-      rsbuildConfig: {
-        output: {
-          injectStyles: true,
-        },
-      },
-    });
-
-    const bundlerConfigs = await rsbuild.initConfigs();
-    expect(bundlerConfigs[0]).toMatchSnapshot();
-  });
-
-  it('should add less-loader with tools.less', async () => {
-    const rsbuild = await createStubRsbuild({
-      plugins: [pluginLess()],
-      rsbuildConfig: {
-        tools: {
-          less: {
-            lessOptions: {
-              javascriptEnabled: false,
-            },
-          },
-        },
-      },
-    });
-
-    const bundlerConfigs = await rsbuild.initConfigs();
-    expect(bundlerConfigs[0]).toMatchSnapshot();
-  });
-
-  it('should add less-loader with excludes', async () => {
-    const rsbuild = await createStubRsbuild({
-      plugins: [pluginLess()],
-      rsbuildConfig: {
-        tools: {
-          less(_config, { addExcludes }) {
-            addExcludes(/node_modules/);
-          },
-        },
-      },
-    });
-
-    const bundlerConfigs = await rsbuild.initConfigs();
     expect(bundlerConfigs[0]).toMatchSnapshot();
   });
 });
