+++ packages/plugin-less/tests/css.test.ts
@@ -1,0 +1,66 @@
+import { createStubRsbuild } from '@scripts/test-helper';
+import { pluginLess } from '../src';
+
+describe('plugin-less', () => {
+  it('should add less-loader', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginLess()],
+      rsbuildConfig: {
+        tools: {
+          less: {},
+        },
+      },
+    });
+
+    const bundlerConfigs = await rsbuild.initConfigs();
+    expect(bundlerConfigs[0]).toMatchSnapshot();
+  });
+
+  it('should add less-loader and css-loader when injectStyles', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginLess()],
+      rsbuildConfig: {
+        output: {
+          injectStyles: true,
+        },
+      },
+    });
+
+    const bundlerConfigs = await rsbuild.initConfigs();
+    expect(bundlerConfigs[0]).toMatchSnapshot();
+  });
+
+  it('should add less-loader with tools.less', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginLess()],
+      rsbuildConfig: {
+        tools: {
+          less: {
+            lessOptions: {
+              javascriptEnabled: false,
+            },
+          },
+        },
+      },
+    });
+
+    const bundlerConfigs = await rsbuild.initConfigs();
+    expect(bundlerConfigs[0]).toMatchSnapshot();
+  });
+
+  it('should add less-loader with excludes', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginLess()],
+      rsbuildConfig: {
+        tools: {
+          less(_config, { addExcludes }) {
+            addExcludes(/node_modules/);
+          },
+        },
+      },
+    });
+
+    const bundlerConfigs = await rsbuild.initConfigs();
+    expect(bundlerConfigs[0]).toMatchSnapshot();
+  });
+});
