+++ packages/plugin-stylus/package.json
@@ -33,10 +33,8 @@
   },
   "devDependencies": {
     "@rsbuild/core": "workspace:*",
-    "@rsbuild/webpack": "workspace:*",
     "@scripts/test-helper": "workspace:*",
-    "typescript": "^5.4.2",
-    "webpack": "^5.91.0"
+    "typescript": "^5.4.2"
   },
   "peerDependencies": {
     "@rsbuild/core": "workspace:^0.7.0-beta.6"
